shisho-server/src/requirements.txt
Flask==3.0.0
transformers
torch
mplfinance
plotly==5.11.0
matplotlib==3.5.2
numpy==1.22.3
pandas==1.4.2
python-binance==1.0.19